Duplicate Logging for cleaner import paths and package separation
