+++ provider/cmd/pulumi-resource-command/schema.json
@@ -129,6 +129,27 @@
         "host"
       ]
     },
+    "command:remote:Logging": {
+      "type": "string",
+      "enum": [
+        {
+          "description": "Capture stdout in logs but not stderr",
+          "value": "stdout"
+        },
+        {
+          "description": "Capture stderr in logs but not stdout",
+          "value": "stderr"
+        },
+        {
+          "description": "Capture stdout and stderr in logs",
+          "value": "stdoutAndStderr"
+        },
+        {
+          "description": "Capture no logs",
+          "value": "none"
+        }
+      ]
+    },
     "command:remote:ProxyConnection": {
       "description": "Instructions for how to connect to a remote endpoint via a bastion host.",
       "properties": {
@@ -368,7 +389,7 @@
           "description": "Additional environment variables available to the command's process.\nNote that this only works if the SSH server is configured to accept these variables via AcceptEnv.\nAlternatively, if a Bash-like shell runs the command on the remote host, you could prefix the command itself\nwith the variables in the form 'VAR=value command'."
         },
         "logging": {
-          "$ref": "#/types/command:local:Logging",
+          "$ref": "#/types/command:remote:Logging",
           "description": "If the command's stdout and stderr should be logged. This doesn't affect the capturing of\nstdout and stderr as outputs. If there might be secrets in the output, you can disable logging here and mark the\noutputs as secret via 'additionalSecretOutputs'. Defaults to logging both stdout and stderr."
         },
         "stderr": {
@@ -424,7 +445,7 @@
           "description": "Additional environment variables available to the command's process.\nNote that this only works if the SSH server is configured to accept these variables via AcceptEnv.\nAlternatively, if a Bash-like shell runs the command on the remote host, you could prefix the command itself\nwith the variables in the form 'VAR=value command'."
         },
         "logging": {
-          "$ref": "#/types/command:local:Logging",
+          "$ref": "#/types/command:remote:Logging",
           "description": "If the command's stdout and stderr should be logged. This doesn't affect the capturing of\nstdout and stderr as outputs. If there might be secrets in the output, you can disable logging here and mark the\noutputs as secret via 'additionalSecretOutputs'. Defaults to logging both stdout and stderr."
         },
         "stdin": {
